+++ src/main/java/com/aps/projectmanage/domain/repository/PermissionRepository.java
@@ -1,10 +1,9 @@
 package com.aps.projectmanage.domain.repository;
 
 import com.aps.projectmanage.domain.entity.Permission;
-import com.aps.projectmanage.domain.entity.Role;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionRepository extends BaseRepository<Permission, Long>{
+public interface PermissionRepository extends BaseRepository<Permission, Integer>{
 
 }
