+++ src/main/java/com/aps/projectmanage/domain/repository/RoleRepository.java
@@ -4,5 +4,5 @@ import com.aps.projectmanage.domain.entity.Role;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends BaseRepository<Role, Long> {
+public interface RoleRepository extends BaseRepository<Role, Integer> {
 }
