+++ src/main/java/com/aps/projectmanage/domain/repository/TaskRepository.java
@@ -4,5 +4,5 @@ import com.aps.projectmanage.domain.entity.Task;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskRepository extends BaseRepository<Task, Long> {
+public interface TaskRepository extends BaseRepository<Task, Integer> {
 }
